Refactor multiple components and reuse across the app. Extract common styles.
