+++ apps/mobile/src/app/(app)/(tabs)/regions/index.tsx
@@ -1,5 +1,5 @@
 import { useState, useCallback } from "react";
-import { View, Pressable, StyleSheet } from "react-native";
+import { Pressable, StyleSheet } from "react-native";
 import { Text } from "react-native-paper";
 import { AddFAB } from "@/components/AddFAB";
 import { SafeAreaView } from "react-native-safe-area-context";
@@ -11,8 +11,9 @@ import { ScreenHeader } from "@/components/ScreenHeader";
 import { useApiQuery } from "@/hooks/use-api-query";
 import { api } from "@/lib/api/client";
 import { queryGate } from "@/lib/functions/query-gate";
+import { commonStyles } from "@/styles/common";
 import { theme } from "@/lib/theme";
-import type { Region, Country } from "@cellarboss/types";
+import type { Region } from "@cellarboss/types";
 
 const SORT_OPTIONS = [
   { label: "Name (A-Z)", value: "name-asc" },
@@ -79,7 +80,7 @@ export default function RegionsScreen() {
   });
 
   return (
-    <SafeAreaView style={styles.container} edges={["top"]}>
+    <SafeAreaView style={commonStyles.screenContainer} edges={["top"]}>
       <ScreenHeader title="Regions" showBack />
       <DataList
         data={sortedRegions}
@@ -116,13 +117,16 @@ export default function RegionsScreen() {
         ]}
         renderItem={(region) => (
           <Pressable
-            style={styles.item}
+            style={commonStyles.listItem}
             onPress={() => router.push(`/regions/${region.id}`)}
           >
-            <Text style={styles.itemTitle} numberOfLines={1}>
+            <Text
+              style={[commonStyles.listItemTitle, styles.title]}
+              numberOfLines={1}
+            >
               {region.name}
             </Text>
-            <Text style={styles.itemSub} numberOfLines={1}>
+            <Text style={commonStyles.listItemSub} numberOfLines={1}>
               {getCountryName(region)}
             </Text>
           </Pressable>
@@ -152,25 +156,7 @@ export default function RegionsScreen() {
 }
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background,
-  },
-  item: {
-    backgroundColor: theme.colors.surface,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    borderBottomWidth: 1,
-    borderBottomColor: theme.colors.outlineVariant,
-  },
-  itemTitle: {
-    fontSize: 15,
-    fontWeight: "bold",
-    color: theme.colors.onSurface,
+  title: {
     marginBottom: 2,
-  },
-  itemSub: {
-    fontSize: 13,
-    color: theme.colors.onSurfaceVariant,
   },
 });
