+++ apps/mobile/src/app/(app)/(tabs)/regions/new.tsx
@@ -1,11 +1,10 @@
-import { StyleSheet } from "react-native";
+import { commonStyles } from "@/styles/common";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { useQueryClient } from "@tanstack/react-query";
 import { useLocalSearchParams } from "expo-router";
 import { ScreenHeader } from "@/components/ScreenHeader";
 import { FormCard } from "@/components/FormCard";
 import { api } from "@/lib/api/client";
-import { theme } from "@/lib/theme";
 import { regionFields } from "@/lib/fields/regions";
 import type { Region } from "@cellarboss/types";
 import type { ApiResult } from "@cellarboss/common";
@@ -39,7 +38,7 @@ export default function NewRegionScreen() {
   };
 
   return (
-    <SafeAreaView style={styles.container} edges={["top"]}>
+    <SafeAreaView style={commonStyles.screenContainer} edges={["top"]}>
       <ScreenHeader title="New Region" showBack />
       <FormCard
         mode="create"
@@ -50,10 +49,3 @@ export default function NewRegionScreen() {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background,
-  },
-});
