+++ apps/mobile/src/app/(app)/(tabs)/tasting-notes/index.tsx
@@ -1,21 +1,19 @@
 import { useState, useCallback } from "react";
-import { View, Pressable, StyleSheet } from "react-native";
-import { Text } from "react-native-paper";
-import { AddFAB } from "@/components/AddFAB";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { useRouter } from "expo-router";
 import { useQueryClient, useMutation } from "@tanstack/react-query";
 import { DataList } from "@/components/DataList";
+import { AddFAB } from "@/components/AddFAB";
 import { ConfirmDialog } from "@/components/ConfirmDialog";
 import { ScreenHeader } from "@/components/ScreenHeader";
+import { TastingNoteListItem } from "@/components/tasting-notes/TastingNoteListItem";
 import { useApiQuery } from "@/hooks/use-api-query";
 import { useSetting } from "@/hooks/use-settings";
 import { api } from "@/lib/api/client";
 import { queryGate } from "@/lib/functions/query-gate";
-import { formatDateTime } from "@/lib/functions/format";
-import { scoreColor } from "@cellarboss/common";
+import { commonStyles } from "@/styles/common";
 import { theme } from "@/lib/theme";
-import type { TastingNote, Vintage, Wine } from "@cellarboss/types";
+import type { TastingNote } from "@cellarboss/types";
 
 const SORT_OPTIONS = [
   { label: "Date (Newest)", value: "date-desc" },
@@ -106,55 +104,53 @@ export default function TastingNotesScreen() {
   });
 
   return (
-    <SafeAreaView style={styles.container} edges={["top"]}>
+    <SafeAreaView style={commonStyles.screenContainer} edges={["top"]}>
       <ScreenHeader title="Tasting Notes" showBack />
-      <View style={styles.content}>
-        <DataList
-          data={sortedNotes}
-          keyExtractor={(item) => String(item.id)}
-          searchPlaceholder="Search tasting notes..."
-          searchFilter={(item, query) => {
-            const lower = query.toLowerCase();
-            const ctx = getWineContext(item);
-            return (
-              ctx.wineTitle.toLowerCase().includes(lower) ||
-              ctx.winemakerName.toLowerCase().includes(lower) ||
-              item.author.toLowerCase().includes(lower) ||
-              item.notes.toLowerCase().includes(lower)
-            );
-          }}
-          sortOptions={SORT_OPTIONS}
-          onSort={setCurrentSort}
-          currentSort={currentSort}
-          onRefresh={handleRefresh}
-          refreshing={refreshing}
-          emptyIcon="note-text-outline"
-          emptyTitle="No tasting notes yet"
-          emptyMessage="Add your first tasting note to get started"
-          emptyActionLabel="Add Tasting Note"
-          onEmptyAction={() => router.push("/tasting-notes/new")}
-          swipeActions={(note) => [
-            {
-              icon: "pencil",
-              color: theme.colors.primary,
-              onPress: () => router.push(`/tasting-notes/${note.id}/edit`),
-            },
-            {
-              icon: "delete",
-              color: "#dc2626",
-              onPress: () => setDeleteTarget(note),
-            },
-          ]}
-          renderItem={(note) => (
-            <TastingNoteListItem
-              note={note}
-              wineContext={getWineContext(note)}
-              datetimeFormat={datetimeFormat}
-              onPress={() => router.push(`/tasting-notes/${note.id}`)}
-            />
-          )}
-        />
-      </View>
+      <DataList
+        data={sortedNotes}
+        keyExtractor={(item) => String(item.id)}
+        searchPlaceholder="Search tasting notes..."
+        searchFilter={(item, query) => {
+          const lower = query.toLowerCase();
+          const ctx = getWineContext(item);
+          return (
+            ctx.wineTitle.toLowerCase().includes(lower) ||
+            ctx.winemakerName.toLowerCase().includes(lower) ||
+            item.author.toLowerCase().includes(lower) ||
+            item.notes.toLowerCase().includes(lower)
+          );
+        }}
+        sortOptions={SORT_OPTIONS}
+        onSort={setCurrentSort}
+        currentSort={currentSort}
+        onRefresh={handleRefresh}
+        refreshing={refreshing}
+        emptyIcon="note-text-outline"
+        emptyTitle="No tasting notes yet"
+        emptyMessage="Add your first tasting note to get started"
+        emptyActionLabel="Add Tasting Note"
+        onEmptyAction={() => router.push("/tasting-notes/new")}
+        swipeActions={(note) => [
+          {
+            icon: "pencil",
+            color: theme.colors.primary,
+            onPress: () => router.push(`/tasting-notes/${note.id}/edit`),
+          },
+          {
+            icon: "delete",
+            color: "#dc2626",
+            onPress: () => setDeleteTarget(note),
+          },
+        ]}
+        renderItem={(note) => (
+          <TastingNoteListItem
+            note={note}
+            wineContext={getWineContext(note)}
+            datetimeFormat={datetimeFormat}
+            onPress={() => router.push(`/tasting-notes/${note.id}`)}
+          />
+        )}
+      />
 
       <AddFAB onPress={() => router.push("/tasting-notes/new")} />
 
@@ -177,119 +173,3 @@ export default function TastingNotesScreen() {
     </SafeAreaView>
   );
 }
-
-type TastingNoteListItemProps = {
-  note: TastingNote;
-  wineContext: { wineTitle: string; winemakerName: string };
-  datetimeFormat?: string | number | boolean | null;
-  onPress: () => void;
-};
-
-function TastingNoteListItem({
-  note,
-  wineContext,
-  datetimeFormat,
-  onPress,
-}: TastingNoteListItemProps) {
-  return (
-    <Pressable style={styles.item} onPress={onPress}>
-      <View
-        style={[styles.scoreBadge, { backgroundColor: scoreColor(note.score) }]}
-      >
-        <Text style={styles.scoreText}>{note.score}</Text>
-      </View>
-      <View style={styles.itemContent}>
-        <View style={styles.itemTop}>
-          <Text style={styles.itemTitle} numberOfLines={1}>
-            {wineContext.wineTitle}
-          </Text>
-          {!!wineContext.winemakerName && (
-            <Text style={styles.winemakerLabel} numberOfLines={1}>
-              {wineContext.winemakerName}
-            </Text>
-          )}
-        </View>
-        {!!note.notes && (
-          <Text style={styles.itemSub} numberOfLines={2}>
-            {note.notes}
-          </Text>
-        )}
-        <View style={styles.itemMeta}>
-          <Text style={styles.metaText}>{note.author}</Text>
-          <Text style={styles.metaText}>
-            {typeof datetimeFormat === "string"
-              ? formatDateTime(note.date, datetimeFormat)
-              : note.date.split("T")[0]}
-          </Text>
-        </View>
-      </View>
-    </Pressable>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background,
-  },
-  content: {
-    flex: 1,
-  },
-  item: {
-    backgroundColor: theme.colors.surface,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    borderBottomWidth: 1,
-    borderBottomColor: theme.colors.outlineVariant,
-    flexDirection: "row",
-    alignItems: "flex-start",
-    gap: 12,
-  },
-  scoreBadge: {
-    width: 36,
-    height: 36,
-    borderRadius: 18,
-    alignItems: "center",
-    justifyContent: "center",
-    marginTop: 2,
-    flexShrink: 0,
-  },
-  scoreText: {
-    fontSize: 14,
-    fontWeight: "bold",
-    color: "#fff",
-  },
-  itemContent: {
-    flex: 1,
-    gap: 3,
-  },
-  itemTop: {
-    flexDirection: "row",
-    alignItems: "center",
-    gap: 8,
-  },
-  itemTitle: {
-    flex: 1,
-    fontSize: 15,
-    fontWeight: "bold",
-    color: theme.colors.onSurface,
-  },
-  winemakerLabel: {
-    fontSize: 13,
-    color: theme.colors.onSurfaceVariant,
-  },
-  itemSub: {
-    fontSize: 13,
-    color: theme.colors.onSurface,
-    lineHeight: 18,
-  },
-  itemMeta: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    marginTop: 2,
-  },
-  metaText: {
-    fontSize: 12,
-    color: theme.colors.onSurfaceVariant,
-  },
-});
