+++ apps/mobile/src/app/(app)/(tabs)/vintages/index.tsx
@@ -1,23 +1,18 @@
 import { useState, useCallback } from "react";
-import { View, Pressable, StyleSheet } from "react-native";
-import { Text } from "react-native-paper";
-import { AddFAB } from "@/components/AddFAB";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { useRouter } from "expo-router";
 import { useQueryClient, useMutation } from "@tanstack/react-query";
 import { DataList } from "@/components/DataList";
+import { AddFAB } from "@/components/AddFAB";
 import { ConfirmDialog } from "@/components/ConfirmDialog";
 import { ScreenHeader } from "@/components/ScreenHeader";
+import { VintageListItem } from "@/components/vintage/VintageListItem";
 import { useApiQuery } from "@/hooks/use-api-query";
 import { api } from "@/lib/api/client";
 import { queryGate } from "@/lib/functions/query-gate";
-import {
-  formatDrinkingWindow,
-  formatDrinkingStatus,
-  type DrinkingStatus,
-} from "@/lib/functions/format";
+import { commonStyles } from "@/styles/common";
 import { theme } from "@/lib/theme";
-import type { Vintage, Wine } from "@cellarboss/types";
+import type { Vintage } from "@cellarboss/types";
 
 const SORT_OPTIONS = [
   { label: "Year (Newest)", value: "year-desc" },
@@ -25,13 +20,6 @@ const SORT_OPTIONS = [
   { label: "Wine (A-Z)", value: "wine-asc" },
   { label: "Wine (Z-A)", value: "wine-desc" },
 ];
-
-const STATUS_COLORS: Record<DrinkingStatus, string> = {
-  drinkable: "#16a34a",
-  wait: "#ca8a04",
-  past: "#dc2626",
-  unknown: theme.colors.onSurfaceVariant,
-};
 
 export default function VintagesScreen() {
   const router = useRouter();
@@ -96,52 +84,50 @@ export default function VintagesScreen() {
   });
 
   return (
-    <SafeAreaView style={styles.container} edges={["top"]}>
+    <SafeAreaView style={commonStyles.screenContainer} edges={["top"]}>
       <ScreenHeader title="Vintages" showBack />
-      <View style={styles.content}>
-        <DataList
-          data={sortedVintages}
-          keyExtractor={(item) => String(item.id)}
-          searchPlaceholder="Search vintages..."
-          searchFilter={(item, query) => {
-            const lower = query.toLowerCase();
-            return (
-              getWineName(item).toLowerCase().includes(lower) ||
-              getYearDisplay(item).toLowerCase().includes(lower)
-            );
-          }}
-          sortOptions={SORT_OPTIONS}
-          onSort={setCurrentSort}
-          currentSort={currentSort}
-          onRefresh={handleRefresh}
-          refreshing={refreshing}
-          emptyIcon="calendar-range"
-          emptyTitle="No vintages yet"
-          emptyMessage="Add your first vintage to get started"
-          emptyActionLabel="Add Vintage"
-          onEmptyAction={() => router.push("/vintages/new")}
-          swipeActions={(vintage) => [
-            {
-              icon: "pencil",
-              color: theme.colors.primary,
-              onPress: () => router.push(`/vintages/${vintage.id}/edit`),
-            },
-            {
-              icon: "delete",
-              color: "#dc2626",
-              onPress: () => setDeleteTarget(vintage),
-            },
-          ]}
-          renderItem={(vintage) => (
-            <VintageListItem
-              vintage={vintage}
-              wineName={getWineName(vintage)}
-              currentYear={currentYear}
-              onPress={() => router.push(`/vintages/${vintage.id}`)}
-            />
-          )}
-        />
-      </View>
+      <DataList
+        data={sortedVintages}
+        keyExtractor={(item) => String(item.id)}
+        searchPlaceholder="Search vintages..."
+        searchFilter={(item, query) => {
+          const lower = query.toLowerCase();
+          return (
+            getWineName(item).toLowerCase().includes(lower) ||
+            getYearDisplay(item).toLowerCase().includes(lower)
+          );
+        }}
+        sortOptions={SORT_OPTIONS}
+        onSort={setCurrentSort}
+        currentSort={currentSort}
+        onRefresh={handleRefresh}
+        refreshing={refreshing}
+        emptyIcon="calendar-range"
+        emptyTitle="No vintages yet"
+        emptyMessage="Add your first vintage to get started"
+        emptyActionLabel="Add Vintage"
+        onEmptyAction={() => router.push("/vintages/new")}
+        swipeActions={(vintage) => [
+          {
+            icon: "pencil",
+            color: theme.colors.primary,
+            onPress: () => router.push(`/vintages/${vintage.id}/edit`),
+          },
+          {
+            icon: "delete",
+            color: "#dc2626",
+            onPress: () => setDeleteTarget(vintage),
+          },
+        ]}
+        renderItem={(vintage) => (
+          <VintageListItem
+            vintage={vintage}
+            wineName={getWineName(vintage)}
+            currentYear={currentYear}
+            onPress={() => router.push(`/vintages/${vintage.id}`)}
+          />
+        )}
+      />
 
       <AddFAB onPress={() => router.push("/vintages/new")} />
 
@@ -164,85 +150,3 @@ export default function VintagesScreen() {
     </SafeAreaView>
   );
 }
-
-type VintageListItemProps = {
-  vintage: Vintage;
-  wineName: string;
-  currentYear: number;
-  onPress: () => void;
-};
-
-function VintageListItem({
-  vintage,
-  wineName,
-  currentYear,
-  onPress,
-}: VintageListItemProps) {
-  const drinkingWindow = formatDrinkingWindow(
-    vintage.drinkFrom,
-    vintage.drinkUntil,
-  );
-  const status = formatDrinkingStatus(
-    vintage.drinkFrom,
-    vintage.drinkUntil,
-    currentYear,
-  );
-
-  return (
-    <Pressable style={styles.item} onPress={onPress}>
-      <View style={styles.itemTop}>
-        <Text style={styles.itemTitle} numberOfLines={1}>
-          {wineName}
-        </Text>
-        <Text style={styles.yearLabel}>
-          {vintage.year !== null ? vintage.year : "NV"}
-        </Text>
-      </View>
-      <Text
-        style={[styles.itemSub, { color: STATUS_COLORS[status] }]}
-        numberOfLines={1}
-      >
-        {drinkingWindow}
-      </Text>
-    </Pressable>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background,
-  },
-  content: {
-    flex: 1,
-  },
-  item: {
-    backgroundColor: theme.colors.surface,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    borderBottomWidth: 1,
-    borderBottomColor: theme.colors.outlineVariant,
-  },
-  itemTop: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    gap: 8,
-    marginBottom: 4,
-  },
-  itemTitle: {
-    flex: 1,
-    fontSize: 15,
-    fontWeight: "bold",
-    color: theme.colors.onSurface,
-  },
-  yearLabel: {
-    fontSize: 14,
-    fontWeight: "600",
-    color: theme.colors.onSurfaceVariant,
-  },
-  itemSub: {
-    fontSize: 13,
-    color: theme.colors.onSurfaceVariant,
-  },
-});
