+++ apps/mobile/src/app/(app)/(tabs)/vintages/new.tsx
@@ -1,11 +1,10 @@
-import { StyleSheet } from "react-native";
+import { commonStyles } from "@/styles/common";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { useLocalSearchParams } from "expo-router";
 import { useQueryClient } from "@tanstack/react-query";
 import { ScreenHeader } from "@/components/ScreenHeader";
 import { FormCard } from "@/components/FormCard";
 import { api } from "@/lib/api/client";
-import { theme } from "@/lib/theme";
 import { vintageFields } from "@/lib/fields/vintages";
 import type { Vintage } from "@cellarboss/types";
 import type { ApiResult } from "@cellarboss/common";
@@ -40,7 +39,7 @@ export default function NewVintageScreen() {
   };
 
   return (
-    <SafeAreaView style={styles.container} edges={["top"]}>
+    <SafeAreaView style={commonStyles.screenContainer} edges={["top"]}>
       <ScreenHeader title="New Vintage" showBack />
       <FormCard
         mode="create"
@@ -51,10 +50,3 @@ export default function NewVintageScreen() {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background,
-  },
-});
