+++ apps/mobile/src/app/(app)/(tabs)/winemakers/[id].tsx
@@ -1,11 +1,11 @@
-import { ScrollView, StyleSheet } from "react-native";
+import { ScrollView } from "react-native";
+import { commonStyles } from "@/styles/common";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { useLocalSearchParams, useRouter } from "expo-router";
 import { ScreenHeader } from "@/components/ScreenHeader";
 import { useApiQuery } from "@/hooks/use-api-query";
 import { api } from "@/lib/api/client";
 import { queryGate } from "@/lib/functions/query-gate";
-import { theme } from "@/lib/theme";
 import { WinemakerDetailsCard } from "@/components/winemaker/WinemakerDetailsCard";
 import { WinemakerWinesList } from "@/components/winemaker/WinemakerWinesList";
 
@@ -24,7 +24,7 @@ export default function ViewWinemakerScreen() {
   const [winemaker] = result.data;
 
   return (
-    <SafeAreaView style={styles.container} edges={["top"]}>
+    <SafeAreaView style={commonStyles.screenContainer} edges={["top"]}>
       <ScreenHeader
         title={winemaker.name}
         showBack
@@ -35,20 +35,10 @@ export default function ViewWinemakerScreen() {
           },
         ]}
       />
-      <ScrollView contentContainerStyle={styles.scroll}>
+      <ScrollView contentContainerStyle={commonStyles.detailScrollContent}>
         <WinemakerDetailsCard winemaker={winemaker} />
         <WinemakerWinesList winemakerId={winemaker.id} />
       </ScrollView>
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background,
-  },
-  scroll: {
-    padding: 16,
-  },
-});
