+++ apps/mobile/src/app/(app)/(tabs)/winemakers/[id]/edit.tsx
@@ -1,4 +1,4 @@
-import { StyleSheet } from "react-native";
+import { commonStyles } from "@/styles/common";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { useLocalSearchParams } from "expo-router";
 import { useQueryClient } from "@tanstack/react-query";
@@ -7,7 +7,6 @@ import { FormCard } from "@/components/FormCard";
 import { useApiQuery } from "@/hooks/use-api-query";
 import { api } from "@/lib/api/client";
 import { queryGate } from "@/lib/functions/query-gate";
-import { theme } from "@/lib/theme";
 import { winemakerFields } from "@/lib/fields/winemakers";
 import type { WineMaker } from "@cellarboss/types";
 import type { ApiResult } from "@cellarboss/common";
@@ -42,7 +41,7 @@ export default function EditWinemakerScreen() {
   };
 
   return (
-    <SafeAreaView style={styles.container} edges={["top"]}>
+    <SafeAreaView style={commonStyles.screenContainer} edges={["top"]}>
       <ScreenHeader title="Edit Winemaker" showBack />
       <FormCard
         mode="edit"
@@ -53,10 +52,3 @@ export default function EditWinemakerScreen() {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background,
-  },
-});
