+++ apps/mobile/src/app/(app)/(tabs)/wines/new.tsx
@@ -1,11 +1,10 @@
-import { StyleSheet } from "react-native";
+import { commonStyles } from "@/styles/common";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { useQueryClient } from "@tanstack/react-query";
 import { useLocalSearchParams } from "expo-router";
 import { ScreenHeader } from "@/components/ScreenHeader";
 import { FormCard } from "@/components/FormCard";
 import { api } from "@/lib/api/client";
-import { theme } from "@/lib/theme";
 import { wineFields, type WineFormData } from "@/lib/fields/wines";
 import type { Wine } from "@cellarboss/types";
 import type { ApiResult } from "@cellarboss/common";
@@ -64,7 +63,7 @@ export default function NewWineScreen() {
   };
 
   return (
-    <SafeAreaView style={styles.container} edges={["top"]}>
+    <SafeAreaView style={commonStyles.screenContainer} edges={["top"]}>
       <ScreenHeader title="New Wine" showBack />
       <FormCard
         mode="create"
@@ -75,10 +74,3 @@ export default function NewWineScreen() {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background,
-  },
-});
